requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyautogen>=0.2.0
supabase>=2.0.0
openai
python-multipart>=0.0.6
pypdf>=3.17.0
tiktoken>=0.5.1
langchain>=0.0.350
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
tenacity>=8.2.3
httpx>=0.25.0

